auth/requirements.txt
Flask==3.0.0
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.2.0
google-api-python-client==2.97.0
httplib2==0.22.0
